+++ github-usercard/src/App.js
@@ -1,15 +1,14 @@
 import React from 'react';
-import UserCard from "./components/UserCard"
+import User from "./components/User"
 import logo from './logo.svg';
 import './App.css';
 
 export default class App extends React.Component {
-
   render(){
     return (
       <div className="App">
         <h1>My Followers</h1>
-        <UserCard />
+        <User />
       </div>
     );
   }
